bugfix: properly remove remaining events of terminating experiment
Before this commit, EventList::getEventsOf() broke when *two* experiment
flows terminated in direct succession: (*it)->getParent() fails if the
underlying event doesn't exist anymore.

To fix this, I moved the event removal from SimulatorController to
EventList::remove(flow), and made it work on the internal event list
magic similarly to EventList::remove(event).

And I tested it with a few known-working experiments.

git-svn-id: https://www4.informatik.uni-erlangen.de/i4svn/danceos/trunk/devel/fail@1022 8c4709b5-6ec9-48aa-a5cd-a96041d1645a