reverting r1671, it completely breaks linking everywhere
FIXME: We really need to make linking less fragile.  The campaign server
should not (at the moment: cannot, and does not) link against the
simulator backend, but the SAL instance refers to backend symbols.  This
just seems to work currently due to lucky translation unit arrangement and
a lot of karma accumulated over the last few years.

git-svn-id: https://www4.informatik.uni-erlangen.de/i4svn/danceos/trunk/devel/fail@1693 8c4709b5-6ec9-48aa-a5cd-a96041d1645a