core/sal: dependency checks moved to listener classes
It is no longer required (and recommended) to write a dependency
check (using conditional compilation) in the experiment code. Instead,
the Fail code performs these checks itself. If an experiment is
compiled without having the required configuration flags set to
enabled, the compiler emits an error. This is because the constructors
of unused listener classes are declared as protected. (Note that
this works only because the Fail code does not create any listeners.)

In a next step, one should remove all the dependency guards in the
existing experiments (not needed anymore).

Change-Id: I1331813a432a752322d510b4102da53ff53294f0