gem5 buildsys: compiler flag fixed due to newer gcc version
(Debian 7.1 comes with gcc 4.7.2) However, we are faking a gcc v4.4.5
but the gem5 build system only adds the '-Wno-unused-but-set-variable'
flag in case of a gcc >= v4.6. Now, this flag will always be appended.
(Older gcc versions should ignore the flag or emit a warning but will
continue to run.)

Change-Id: I7021e08019951375c7e8b56d0f875191f904971d