util/llvmdisassembler: build with ac++-clang
This change updates the documentation on how to manually build LLVM
3.3 or 3.4, and removes the forced -fno-rtti compiler flag from
util/llvmdisassembler/: The flag breaks compilation with (not yet
released) AspectC++ 2.0, which complains about dynamic_cast (used in
some active aspect headers that should not affect llvmdisassembler at
all) being unusable with -fno-rtti.  It's probably not a good idea to
compile only a part of Fail* with -fno-rtti anyways.

Note that the Debian and Ubuntu LLVM packages are built with
REQUIRES_RTTI=1, too.

Change-Id: I9891cf074d4201df786f7f5a9b96033e18832562